+++ app/(gcforms)/[locale]/(form filler)/id/[...props]/clientSide.tsx
@@ -10,10 +10,12 @@ export const FormWrapper = ({
   formRecord,
   currentForm,
   allowGrouping,
+  clientIp,
 }: {
   formRecord: TypeOmit<FormRecord, "name" | "deliveryOption">;
   currentForm: JSX.Element[];
   allowGrouping?: boolean | undefined;
+  clientIp: string;
 }) => {
   // TODO cast language as "en" | "fr" in TS below
   const {
@@ -41,6 +43,7 @@ export const FormWrapper = ({
         );
       }}
       allowGrouping={allowGrouping}
+      clientIp={clientIp}
     >
       {currentForm}
     </Form>
